requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altgraph==0.17.4
asyncio==3.4.3
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
colorama==0.4.6
discord-protos==0.0.2
discord.py-self @ git+https://github.com/pc111c/discord.py-self.git@e16f555c9e78411b1511fee2273448083d07ab63
frozenlist==1.4.1
idna==3.7
logging==0.4.9.6
multidict==6.0.5
packaging==24.1
pefile==2024.8.26
protobuf==5.27.1
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
pywin32-ctypes==0.2.3
requests==2.32.3
setuptools==75.1.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2
yarl==1.9.4
